requirements.txt
mlflow
boto3
scikit-learn
python-dotenv
pandas
numpy
fastapi
uvicorn
